+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ v5.27.0
 
 * Point LOVE weather station to read corresponding CSC `<https://github.com/lsst-ts/LOVE-frontend/pull/553>`_
 * Add Dynalene Component `<https://github.com/lsst-ts/LOVE-frontend/pull/552>`_
+* Add new params: category and time_lost_type to narrative log `<https://github.com/lsst-ts/LOVE-frontend/pull/551>`_
 
 v5.26.1
 -------